:truck: collection data 상태관리 폴더 분리 #62
