+++ src/features/collection/api.ts
@@ -1,0 +1,9 @@
+import { client } from '@/features/api';
+
+export const getCollections = async () => {
+  return client.get<Collections>('/collection').then((response) => {
+    const { data } = response;
+    console.debug(data);
+    return data;
+  });
+};
