+++ src/features/collection/constants.ts
@@ -1,0 +1,1 @@
+export const COLLECTION_KEY = 'collection';
