+++ src/features/collection/index.ts
@@ -1,0 +1,2 @@
+export { refetchCollection } from './utils';
+export * from './hooks';
