+++ src/features/collection/utils.ts
@@ -1,0 +1,8 @@
+import { COLLECTION_KEY } from './constants';
+
+import { queryClient } from '@/features/api';
+
+export const refetchCollection = () =>
+  queryClient.invalidateQueries({
+    queryKey: [COLLECTION_KEY],
+  });
